backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
django-filter>=24.2
psycopg[binary]>=3.1
django-cors-headers>=4.3
drf-spectacular>=0.27
django-import-export>=4.1
Pillow>=10.0
cryptography>=42.0
Markdown>=3.6
reportlab>=4.0
bleach>=6.1
gunicorn>=22.0
django-redis>=5.4
python-magic>=0.4

# testing
pytest>=8.0
pytest-django>=4.8

# dev / CI tooling
mypy>=1.10
types-bleach>=6.1
